chore: Rename folder to [lessonSlug]
